refactor(cli): collapse param/secret show/diff/list/log/tag/untag into generic commands
Both usecase trees are on the provider interface (#202/#203), so the
near-identical param/secret CLI twins for show/diff/list/log/tag/untag
collapse into one generic command each. Behavior-preserving: CLI output
byte-identical (golden tests unedited).

Design: generic command scaffold (arg/flag/spec dispatch, pager,
raw/json/text dispatch, Runner shape, error handling) under
internal/cli/commands/generic/**, parameterized by a small per-provider
Config + Presenter. The provider-specific Presenter reproduces each
provider's exact output (param: Version/Type-via-paramtype/Modified;
secret: ARN-from-Entry.Extra/VersionId/Stages/Created). The 12 twin
subpackages are deleted; create/update/delete/restore stay per-provider
(provider-specific options; out of scope).

- log: 140-line Run split into per-mode renderers (default/patch/oneline/
  json); //nolint:gocognit removed.
- tag/untag: unified over provider.Tagger; duplicated parseTags is one
  helper; //nolint:dupl removed.
- Tests: consolidated into shared golden harnesses running BOTH providers
  through the generic Runner + presenter, reusing providermock. Every
  asserted expected-output preserved verbatim. Test LOC 3642 -> 2949.

Note: prod LOC is ~flat (not the audit's ~1000->350). Because the two
providers render structurally-different metadata and keep distinct golden
help/flags, the generic-scaffold + thin-presenter seam is LOC-neutral at
two providers; the structural duplication (twin Run bodies, action flow,
parseTags, log branching) is eliminated and the seam amortizes as GCP/
Azure providers (#207/#208) are added. It also unblocks #205.

Closes #204
