+++ internal/cli/commands/param/command.go
@@ -7,12 +7,6 @@ import (
 	cliinternal "github.com/mpyw/suve/internal/cli/commands/internal"
 	"github.com/mpyw/suve/internal/cli/commands/param/create"
 	paramdelete "github.com/mpyw/suve/internal/cli/commands/param/delete"
-	"github.com/mpyw/suve/internal/cli/commands/param/diff"
-	"github.com/mpyw/suve/internal/cli/commands/param/list"
-	"github.com/mpyw/suve/internal/cli/commands/param/log"
-	"github.com/mpyw/suve/internal/cli/commands/param/show"
-	"github.com/mpyw/suve/internal/cli/commands/param/tag"
-	"github.com/mpyw/suve/internal/cli/commands/param/untag"
 	"github.com/mpyw/suve/internal/cli/commands/param/update"
 )
 
@@ -23,15 +17,15 @@ func Command() *cli.Command {
 		Aliases: []string{"ssm", "ps"},
 		Usage:   "Interact with AWS Systems Manager Parameter Store",
 		Commands: []*cli.Command{
-			show.Command(),
-			log.Command(),
-			diff.Command(),
-			list.Command(),
+			ShowCommand(),
+			LogCommand(),
+			DiffCommand(),
+			ListCommand(),
 			create.Command(),
 			update.Command(),
 			paramdelete.Command(),
-			tag.Command(),
-			untag.Command(),
+			TagCommand(),
+			UntagCommand(),
 		},
 		CommandNotFound: cliinternal.CommandNotFound,
 	}
