--- internal/cli/commands/param/diff/diff_test.go
@@ -1,540 +1,0 @@
-package diff_test
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"strings"
-	"testing"
-
-	"github.com/samber/lo"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	appcli "github.com/mpyw/suve/internal/cli/commands"
-	paramdiff "github.com/mpyw/suve/internal/cli/commands/param/diff"
-	"github.com/mpyw/suve/internal/cli/diffargs"
-	"github.com/mpyw/suve/internal/domain"
-	"github.com/mpyw/suve/internal/provider"
-	"github.com/mpyw/suve/internal/provider/providermock"
-	"github.com/mpyw/suve/internal/usecase/param"
-	"github.com/mpyw/suve/internal/version/paramversion"
-)
-
-func TestCommand_Validation(t *testing.T) {
-	t.Parallel()
-
-	t.Run("missing arguments", func(t *testing.T) {
-		t.Parallel()
-
-		app := appcli.MakeApp()
-		err := app.Run(t.Context(), []string{"suve", "param", "diff"})
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "usage:")
-	})
-
-	t.Run("invalid version spec", func(t *testing.T) {
-		t.Parallel()
-
-		app := appcli.MakeApp()
-		err := app.Run(t.Context(), []string{"suve", "param", "diff", "/app/param#"})
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "must be followed by")
-	})
-
-	t.Run("too many arguments", func(t *testing.T) {
-		t.Parallel()
-
-		app := appcli.MakeApp()
-		err := app.Run(t.Context(), []string{"suve", "param", "diff", "/a", "#1", "#2", "#3"})
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "usage:")
-	})
-}
-
-//nolint:funlen // Table-driven test with many cases
-func TestParseArgs(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		args       []string
-		wantSpec1  *wantSpec
-		wantSpec2  *wantSpec
-		wantErrMsg string
-	}{
-		// === 1 argument: full spec vs latest ===
-		{
-			name: "1 arg: version specified",
-			args: []string{"/app/config#3"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(3)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-		{
-			name: "1 arg: shift specified",
-			args: []string{"/app/config~1"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   1,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-		{
-			name: "1 arg: version and shift",
-			args: []string{"/app/config#5~2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(5)),
-				shift:   2,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-		{
-			name: "1 arg: no version (latest vs latest)",
-			args: []string{"/app/config"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-
-		// === 2 arguments: second starts with # or ~ (use name from first) ===
-		{
-			name: "2 args: name + version spec (partial spec)",
-			args: []string{"/app/config", "#3"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(3)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: full spec + version spec (mixed)",
-			args: []string{"/app/config#1", "#2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(1)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(2)),
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: full spec with shift + version spec",
-			args: []string{"/app/config~1", "#2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   1,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(2)),
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: name + shift spec",
-			args: []string{"/app/config", "~"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   1,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: name + shift spec ~2",
-			args: []string{"/app/config", "~2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   2,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-
-		// === 2 arguments: fullpath x2 ===
-		{
-			name: "2 args: full spec x2 same key",
-			args: []string{"/app/config#1", "/app/config#2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(1)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(2)),
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: full spec x2 different keys",
-			args: []string{"/app/config#1", "/other/key#2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(1)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/other/key",
-				version: lo.ToPtr(int64(2)),
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: first latest, second versioned",
-			args: []string{"/app/config", "/app/config#2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(2)),
-				shift:   0,
-			},
-		},
-		{
-			name: "2 args: first versioned, second latest",
-			args: []string{"/app/config#1", "/app/config"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(1)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   0,
-			},
-		},
-
-		// === 3 arguments: partial spec format ===
-		{
-			name: "3 args: partial spec format",
-			args: []string{"/app/config", "#1", "#2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(1)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(2)),
-				shift:   0,
-			},
-		},
-		{
-			name: "3 args: partial spec with shifts",
-			args: []string{"/app/config", "~1", "~2"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   1,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   2,
-			},
-		},
-		{
-			name: "3 args: partial spec mixed version and shift",
-			args: []string{"/app/config", "#3", "~"},
-			wantSpec1: &wantSpec{
-				name:    "/app/config",
-				version: lo.ToPtr(int64(3)),
-				shift:   0,
-			},
-			wantSpec2: &wantSpec{
-				name:    "/app/config",
-				version: nil,
-				shift:   1,
-			},
-		},
-
-		// === Error cases ===
-		{
-			name:       "0 args: error",
-			args:       []string{},
-			wantErrMsg: "usage:",
-		},
-		{
-			name:       "4+ args: error",
-			args:       []string{"/app/config", "#1", "#2", "#3"},
-			wantErrMsg: "usage:",
-		},
-		{
-			name:       "invalid shift spec in 1 arg",
-			args:       []string{"/app/config#3~abc"},
-			wantErrMsg: "invalid",
-		},
-		{
-			name:       "invalid shift spec in 2nd arg",
-			args:       []string{"/app/config", "#3~abc"},
-			wantErrMsg: "invalid",
-		},
-		{
-			name:       "2 args: invalid first arg",
-			args:       []string{"#", "/app/config#2"},
-			wantErrMsg: "invalid first argument",
-		},
-		{
-			name:       "2 args: invalid second arg (full spec x2)",
-			args:       []string{"/app/config#1", "/app/config#"},
-			wantErrMsg: "invalid second argument",
-		},
-		{
-			name:       "3 args: invalid version1",
-			args:       []string{"/app/config", "#", "#2"},
-			wantErrMsg: "invalid version1",
-		},
-		{
-			name:       "3 args: invalid version2",
-			args:       []string{"/app/config", "#1", "#"},
-			wantErrMsg: "invalid version2",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			spec1, spec2, err := diffargs.ParseArgs(
-				tt.args,
-				paramversion.Parse,
-				func(abs paramversion.AbsoluteSpec) bool { return abs.Version != nil },
-				"#~",
-				"usage: suve param diff <spec1> [spec2] | <name> <version1> [version2]",
-			)
-
-			if tt.wantErrMsg != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.wantErrMsg)
-
-				return
-			}
-
-			require.NoError(t, err)
-			assertSpec(t, "spec1", spec1, tt.wantSpec1)
-			assertSpec(t, "spec2", spec2, tt.wantSpec2)
-		})
-	}
-}
-
-type wantSpec struct {
-	name    string
-	version *int64
-	shift   int
-}
-
-func assertSpec(t *testing.T, label string, got *paramversion.Spec, want *wantSpec) {
-	t.Helper()
-	assert.Equal(t, want.name, got.Name, "%s.Name", label)
-	assert.Equal(t, want.version, got.Absolute.Version, "%s.Absolute.Version", label)
-	assert.Equal(t, want.shift, got.Shift, "%s.Shift", label)
-}
-
-// diffStore resolves a version-spec suffix ("#N" or "" for latest) to a ref and
-// returns the mapped entry, erroring for unmapped refs (simulating not-found).
-func diffStore(byRef map[string]*domain.Entry) *providermock.Store {
-	return &providermock.Store{
-		ResolveFunc: func(_ context.Context, _, spec string) (provider.VersionRef, error) {
-			return provider.NewVersionRef(strings.TrimPrefix(spec, "#")), nil
-		},
-		GetFunc: func(_ context.Context, _ string, ref provider.VersionRef) (*domain.Entry, error) {
-			entry, ok := byRef[ref.ID()]
-			if !ok {
-				return nil, fmt.Errorf("version not found")
-			}
-
-			return entry, nil
-		},
-	}
-}
-
-func versionSpec(v int64) *paramversion.Spec {
-	return &paramversion.Spec{Name: "/app/param", Absolute: paramversion.AbsoluteSpec{Version: lo.ToPtr(v)}}
-}
-
-func TestRun(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		opts    paramdiff.Options
-		store   *providermock.Store
-		wantErr bool
-		check   func(t *testing.T, output string)
-	}{
-		{
-			name: "diff between two versions",
-			opts: paramdiff.Options{Spec1: versionSpec(1), Spec2: versionSpec(2)},
-			store: diffStore(map[string]*domain.Entry{
-				"1": {Name: "/app/param", Value: "old-value", Version: domain.Version{ID: "1"}},
-				"2": {Name: "/app/param", Value: "new-value", Version: domain.Version{ID: "2"}},
-			}),
-			check: func(t *testing.T, output string) {
-				t.Helper()
-				assert.Contains(t, output, "-old-value")
-				assert.Contains(t, output, "+new-value")
-			},
-		},
-		{
-			name: "no diff when same content",
-			opts: paramdiff.Options{Spec1: versionSpec(1), Spec2: versionSpec(2)},
-			store: diffStore(map[string]*domain.Entry{
-				"1": {Name: "/app/param", Value: "same-value", Version: domain.Version{ID: "1"}},
-				"2": {Name: "/app/param", Value: "same-value", Version: domain.Version{ID: "2"}},
-			}),
-			check: func(t *testing.T, output string) {
-				t.Helper()
-				assert.NotContains(t, output, "-same-value")
-			},
-		},
-		{
-			name: "error getting first version",
-			opts: paramdiff.Options{Spec1: versionSpec(1), Spec2: versionSpec(2)},
-			store: diffStore(map[string]*domain.Entry{
-				"2": {Name: "/app/param", Value: "value", Version: domain.Version{ID: "2"}},
-			}),
-			wantErr: true,
-		},
-		{
-			name: "error getting second version",
-			opts: paramdiff.Options{Spec1: versionSpec(1), Spec2: versionSpec(2)},
-			store: diffStore(map[string]*domain.Entry{
-				"1": {Name: "/app/param", Value: "value", Version: domain.Version{ID: "1"}},
-			}),
-			wantErr: true,
-		},
-		{
-			name: "json format with valid JSON values",
-			opts: paramdiff.Options{Spec1: versionSpec(1), Spec2: versionSpec(2), ParseJSON: true},
-			store: diffStore(map[string]*domain.Entry{
-				"1": {Name: "/app/param", Value: `{"key":"old"}`, Version: domain.Version{ID: "1"}},
-				"2": {Name: "/app/param", Value: `{"key":"new"}`, Version: domain.Version{ID: "2"}},
-			}),
-			check: func(t *testing.T, output string) {
-				t.Helper()
-				assert.Contains(t, output, "-")
-				assert.Contains(t, output, "+")
-			},
-		},
-		{
-			name: "json format with non-JSON values warns",
-			opts: paramdiff.Options{Spec1: versionSpec(1), Spec2: versionSpec(2), ParseJSON: true},
-			store: diffStore(map[string]*domain.Entry{
-				"1": {Name: "/app/param", Value: "not json", Version: domain.Version{ID: "1"}},
-				"2": {Name: "/app/param", Value: "also not json", Version: domain.Version{ID: "2"}},
-			}),
-			check: func(t *testing.T, output string) {
-				t.Helper()
-				assert.Contains(t, output, "-not json")
-				assert.Contains(t, output, "+also not json")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			var buf, errBuf bytes.Buffer
-
-			r := &paramdiff.Runner{
-				UseCase: &param.DiffUseCase{Reader: tt.store},
-				Stdout:  &buf,
-				Stderr:  &errBuf,
-			}
-			err := r.Run(t.Context(), tt.opts)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-
-				return
-			}
-
-			require.NoError(t, err)
-
-			if tt.check != nil {
-				tt.check(t, buf.String())
-			}
-		})
-	}
-}
-
-func TestRun_IdenticalWarning(t *testing.T) {
-	t.Parallel()
-
-	store := diffStore(map[string]*domain.Entry{
-		"": {Name: "/app/param", Value: "same-value", Version: domain.Version{ID: "1"}},
-	})
-
-	var stdout, stderr bytes.Buffer
-
-	r := &paramdiff.Runner{
-		UseCase: &param.DiffUseCase{Reader: store},
-		Stdout:  &stdout,
-		Stderr:  &stderr,
-	}
-	opts := paramdiff.Options{
-		Spec1: &paramversion.Spec{Name: "/app/param", Absolute: paramversion.AbsoluteSpec{}},
-		Spec2: &paramversion.Spec{Name: "/app/param", Absolute: paramversion.AbsoluteSpec{}},
-	}
-
-	err := r.Run(t.Context(), opts)
-	require.NoError(t, err)
-
-	// stdout should be empty (no diff output)
-	assert.Empty(t, stdout.String())
-
-	// stderr should contain warning and hint
-	stderrStr := stderr.String()
-	assert.Contains(t, stderrStr, "Warning:")
-	assert.Contains(t, stderrStr, "comparing identical versions")
-	assert.Contains(t, stderrStr, "Hint:")
-	assert.Contains(t, stderrStr, "/app/param~1")
-}
