+++ internal/cli/commands/secret/command.go
@@ -7,13 +7,7 @@ import (
 	cliinternal "github.com/mpyw/suve/internal/cli/commands/internal"
 	"github.com/mpyw/suve/internal/cli/commands/secret/create"
 	secretdelete "github.com/mpyw/suve/internal/cli/commands/secret/delete"
-	"github.com/mpyw/suve/internal/cli/commands/secret/diff"
-	"github.com/mpyw/suve/internal/cli/commands/secret/list"
-	"github.com/mpyw/suve/internal/cli/commands/secret/log"
 	"github.com/mpyw/suve/internal/cli/commands/secret/restore"
-	"github.com/mpyw/suve/internal/cli/commands/secret/show"
-	"github.com/mpyw/suve/internal/cli/commands/secret/tag"
-	"github.com/mpyw/suve/internal/cli/commands/secret/untag"
 	"github.com/mpyw/suve/internal/cli/commands/secret/update"
 )
 
@@ -24,16 +18,16 @@ func Command() *cli.Command {
 		Aliases: []string{"sm"},
 		Usage:   "Interact with AWS Secrets Manager",
 		Commands: []*cli.Command{
-			show.Command(),
-			log.Command(),
-			diff.Command(),
-			list.Command(),
+			ShowCommand(),
+			LogCommand(),
+			DiffCommand(),
+			ListCommand(),
 			create.Command(),
 			update.Command(),
 			secretdelete.Command(),
 			restore.Command(),
-			tag.Command(),
-			untag.Command(),
+			TagCommand(),
+			UntagCommand(),
 		},
 		CommandNotFound: cliinternal.CommandNotFound,
 	}
